🗃Refactor: 컴포넌트 클래스 생성
